Update to parse booleans
I also refactored a ton of code because it was messy.
I added support for an array of json objects too.
Corrected nested json in arrays errors as well.

+++ json-parser/include/scanner.h
@@ -11,7 +11,7 @@ typedef struct JSONBuffer JSONBuffer;
 
 struct JSONBuffer {
   char* current_file;
-  int current_position;
+  size_t current_position;
   size_t file_size;
   bool error;
 };
